refactor: ajusta uint

+++ src/core/application/dtos/pay_order_dto.go
@@ -1,5 +1,5 @@
 package dtos
 
 type PayOrderDto struct {
-	OrderId int64 `json:"orderId"`
+	OrderId uint `json:"orderId"`
 }
